+++ components/consts/index.ts
@@ -1,0 +1,3 @@
+export * from './Timing';
+
+export const isDev = import.meta.env.MODE === 'development';
