+++ entrypoints/twitch.content/consts/index.ts
@@ -1,6 +1,3 @@
 export * from './MessageTemplates';
 export * from './Commands';
-export * from './Timing';
 export * from './StreamStatus';
-
-export const isDev = import.meta.env.MODE === 'development';
