+++ entrypoints/twitch.content/modules/stream/services/streamStatus/StreamStatusService.ts
@@ -1,5 +1,5 @@
 import { ColorService } from '@twitch/modules/shared';
-import { StreamStatus, Timing } from '@twitch/consts';
+import { StreamStatus } from '@twitch/consts';
 import './style.css';
 import { ChatFacade } from '@twitch/modules/chat';
 import { Container, ContainerInstance } from 'typedi';
@@ -8,6 +8,7 @@ import { OnScreenTextRecognizer } from '@components/services';
 import { getRandomNumber, logDev } from '@components/utils';
 import { BasicView } from '@components/BasicView';
 import { EventEmitter } from '@components/EventEmitter';
+import { Timing } from '@components/consts';
 import {
     antiCheatChecks, anticheatName, chestGameChecks, ICheck, lootGameChecks
 } from './checks';
