+++ entrypoints/twitch.content/views/extensionContainer/composables/useChestGameService.ts
@@ -1,4 +1,4 @@
-import { Timing } from '@twitch/consts';
+import { Timing } from '@components/consts';
 import { StreamFacade } from '@twitch/modules/stream';
 import { LocalSettingsService } from '@components/settings';
 import { ChestGameService } from '@twitch/modules/miniGames';
